Add a "tempDir" option for the locale file output
This is a place to write the temporary locale files that get included
into the webpack bundle. These are now outside of the output path
so that they don't cause the webpack watch mode to trigger a new
re-packing when the plugin updates them at the end of the first
packing.